backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# GIS
geoalchemy2==0.14.2
# shapely==2.0.2  # Опционально, для сложных операций
# pyproj==3.6.1   # Опционально, для проекций

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
PyJWT==2.8.0  # Для JWT токенов (уже используется в коде)

# Cache
redis==5.0.1
# hiredis==2.2.3  # Требует Rust

# Utils
python-dotenv==1.0.0
httpx==0.25.2  # Для отправки SMS через SMSC.RU
aiofiles==23.2.1  # Асинхронная работа с файлами
Pillow==10.1.0  # Обработка изображений
pymysql==1.1.0  # Для миграции данных из MySQL/MariaDB
